Refactor file and directory structure of luac bin plugin
