+++ librz/bin/format/luac/luac_common.c
@@ -1,0 +1,6 @@
+// SPDX-License-Identifier: LGPL-3.0-only
+// SPDX-FileCopyrightText: 2021 Heersin <teablearcher@gmail.com>
+
+// Implement Functions declared in luac_specs.h
+
+#include "luac_common.h"
