+++ librz/bin/meson.build
@@ -101,8 +101,9 @@ rz_bin_sources = [
   'format/mdmp/mdmp_pe.c',
   'format/mdmp/mdmp_pe64.c',
   'format/le/le.c',
-  'format/luac/luac.c',
-  'format/luac/luac_54.c',
+  'format/luac/luac_common.c',
+  'format/luac/v54/bin_54.c',
+  'format/luac/v54/parse_54.c',
   'format/ne/ne.c',
   'format/mz/mz.c',
   'format/nxo/nxo.c',
